refactor(agentic): decompose posting preflight and verification storage modules (#93, #94)
