+++ backend/tests/services/agentic/test_verification_graph.py
@@ -136,7 +136,7 @@ async def test_slice_1_x_profile_fuzzy_match() -> None:
 
     with (
         patch(
-            "app.services.agentic.verification_graph._load_target_posts_from_db",
+            "app.services.agentic.verification_graph.load_target_posts_from_db",
             return_value=[
                 {
                     "id": post_id,
@@ -179,7 +179,7 @@ async def test_slice_2_x_profile_unverified_post() -> None:
 
     with (
         patch(
-            "app.services.agentic.verification_graph._load_target_posts_from_db",
+            "app.services.agentic.verification_graph.load_target_posts_from_db",
             return_value=[
                 {
                     "id": post_id,
@@ -214,7 +214,7 @@ async def test_slice_3_linkedin_post_verification() -> None:
 
     with (
         patch(
-            "app.services.agentic.verification_graph._load_target_posts_from_db",
+            "app.services.agentic.verification_graph.load_target_posts_from_db",
             return_value=[
                 {
                     "id": post_id,
@@ -271,7 +271,7 @@ async def test_slice_4_dual_platform_batch_verification() -> None:
 
     with (
         patch(
-            "app.services.agentic.verification_graph._load_target_posts_from_db",
+            "app.services.agentic.verification_graph.load_target_posts_from_db",
             return_value=target_posts,
         ),
         patch(
@@ -366,7 +366,7 @@ async def test_slice_8_cross_posting_dual_channel_verification() -> None:
 
     with (
         patch(
-            "app.services.agentic.verification_graph._load_target_posts_from_db",
+            "app.services.agentic.verification_graph.load_target_posts_from_db",
             return_value=target_posts,
         ),
         patch(
@@ -400,7 +400,7 @@ async def test_slice_9_empty_target_posts_error_status() -> None:
     missing_id = str(uuid.uuid4())
 
     with patch(
-        "app.services.agentic.verification_graph._load_target_posts_from_db",
+        "app.services.agentic.verification_graph.load_target_posts_from_db",
         return_value=[],
     ):
         report = await verify_posts_with_graph(
